+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # CHANGELOG
 
-## Next Release
+## v6.5.0 (2023-05-09)
 
 - Adds missing typescript definitions for `Utility` class
-- Fixs incorrect typescript definitions for `Address`, `EndShipper`m and `Webhook` classes methods
+- Fixes incorrect typescript definitions for `Address`, `EndShipper` and `Webhook` class methods
 - Moves `Utils` to EasyPostClient class
 
 ## v6.4.0 (2023-05-02)